refactor: move files to `protocol` directory
