+++ zcached/protocol/deserializer.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any, List, Callable, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .reader import Reader
+    from zcached.protocol.reader import Reader
 
 
 class Deserializer:
